+++ shared.h
@@ -3,7 +3,7 @@
 #include <cstring>
 
 #define PROJECT_NAME "WrappEm"
-#define PROJECT_VERSION "v1.1.0"
+#define PROJECT_VERSION "v1.2.0"
 #define PROJECT_LICENSE "MS-RL License"
 #define PROJECT_COPYRIGHT "Copyright (c) 2025-2026. Augusto Goulart."